+++ src/test/ssr.test.js
@@ -1,9 +1,8 @@
 /**
  * @jest-environment node
  */
-
 import React from 'react'
-import { renderToString } from 'react-dom/server'
+import { renderToString, renderToNodeStream } from 'react-dom/server'
 import ServerStyleSheet from '../models/ServerStyleSheet'
 import { resetStyled } from './utils'
 import _injectGlobal from '../constructors/injectGlobal'
@@ -38,7 +37,9 @@ describe('ssr', () => {
     `
 
     const sheet = new ServerStyleSheet()
-    const html = renderToString(sheet.collectStyles(<Heading>Hello SSR!</Heading>))
+    const html = renderToString(
+      sheet.collectStyles(<Heading>Hello SSR!</Heading>)
+    )
     const css = sheet.getStyleTags()
 
     expect(html).toMatchSnapshot()
@@ -54,7 +55,9 @@ describe('ssr', () => {
     `
 
     const sheet = new ServerStyleSheet()
-    const html = renderToString(sheet.collectStyles(<Heading>Hello SSR!</Heading>))
+    const html = renderToString(
+      sheet.collectStyles(<Heading>Hello SSR!</Heading>)
+    )
     const css = sheet.getStyleTags()
 
     expect(html).toMatchSnapshot()
@@ -73,7 +76,9 @@ describe('ssr', () => {
     `
 
     const sheet = new ServerStyleSheet()
-    const html = renderToString(sheet.collectStyles(<Heading>Hello SSR!</Heading>))
+    const html = renderToString(
+      sheet.collectStyles(<Heading>Hello SSR!</Heading>)
+    )
     const css = sheet.getStyleTags()
 
     expect(html).toMatchSnapshot()
@@ -81,20 +86,22 @@ describe('ssr', () => {
   })
 
   it('should render CSS in the order the components were defined, not rendered', () => {
-    const ONE = styled.h1.withConfig({ componentId: 'ONE' }) `
+    const ONE = styled.h1.withConfig({ componentId: 'ONE' })`
       color: red;
     `
-    const TWO = styled.h2.withConfig({ componentId: 'TWO' }) `
+    const TWO = styled.h2.withConfig({ componentId: 'TWO' })`
       color: blue;
     `
 
     const sheet = new ServerStyleSheet()
-    const html = renderToString(sheet.collectStyles(
-      <div>
-        <TWO />
-        <ONE />
-      </div>
-    ))
+    const html = renderToString(
+      sheet.collectStyles(
+        <div>
+          <TWO />
+          <ONE />
+        </div>
+      )
+    )
     const css = sheet.getStyleTags()
 
     expect(html).toMatchSnapshot()
@@ -105,19 +112,23 @@ describe('ssr', () => {
     injectGlobal`
       body { background: papayawhip; }
     `
-    const PageOne = styled.h1.withConfig({ componentId: 'PageOne' }) `
+    const PageOne = styled.h1.withConfig({ componentId: 'PageOne' })`
       color: red;
     `
-    const PageTwo = styled.h2.withConfig({ componentId: 'PageTwo' }) `
+    const PageTwo = styled.h2.withConfig({ componentId: 'PageTwo' })`
       color: blue;
     `
 
     const sheetOne = new ServerStyleSheet()
-    const htmlOne = renderToString(sheetOne.collectStyles(<PageOne>Camera One!</PageOne>))
+    const htmlOne = renderToString(
+      sheetOne.collectStyles(<PageOne>Camera One!</PageOne>)
+    )
     const cssOne = sheetOne.getStyleTags()
 
     const sheetTwo = new ServerStyleSheet()
-    const htmlTwo = renderToString(sheetTwo.collectStyles(<PageTwo>Camera Two!</PageTwo>))
+    const htmlTwo = renderToString(
+      sheetTwo.collectStyles(<PageTwo>Camera Two!</PageTwo>)
+    )
     const cssTwo = sheetTwo.getStyleTags()
 
     expect(htmlOne).toMatchSnapshot()
@@ -128,15 +139,17 @@ describe('ssr', () => {
 
   it('should allow global styles to be injected during rendering', () => {
     injectGlobal`html::before { content: 'Before both renders'; }`
-    const PageOne = styled.h1.withConfig({ componentId: 'PageOne' }) `
+    const PageOne = styled.h1.withConfig({ componentId: 'PageOne' })`
       color: red;
     `
-    const PageTwo = styled.h2.withConfig({ componentId: 'PageTwo' }) `
+    const PageTwo = styled.h2.withConfig({ componentId: 'PageTwo' })`
       color: blue;
     `
 
     const sheetOne = new ServerStyleSheet()
-    const htmlOne = renderToString(sheetOne.collectStyles(<PageOne>Camera One!</PageOne>))
+    const htmlOne = renderToString(
+      sheetOne.collectStyles(<PageOne>Camera One!</PageOne>)
+    )
     injectGlobal`html::before { content: 'During first render'; }`
     const cssOne = sheetOne.getStyleTags()
 
@@ -144,7 +157,9 @@ describe('ssr', () => {
 
     const sheetTwo = new ServerStyleSheet()
     injectGlobal`html::before { content: 'During second render'; }`
-    const htmlTwo = renderToString(sheetTwo.collectStyles(<PageTwo>Camera Two!</PageTwo>))
+    const htmlTwo = renderToString(
+      sheetTwo.collectStyles(<PageTwo>Camera Two!</PageTwo>)
+    )
     const cssTwo = sheetTwo.getStyleTags()
 
     injectGlobal`html::before { content: 'After both renders'; }`
@@ -159,14 +174,14 @@ describe('ssr', () => {
     injectGlobal`
       body { background: papayawhip; }
     `
-    const Header = styled.h1.withConfig({ componentId: 'Header' }) `
-      animation: ${ props => props.animation} 1s both;
+    const Header = styled.h1.withConfig({ componentId: 'Header' })`
+      animation: ${props => props.animation} 1s both;
     `
 
     const sheet = new ServerStyleSheet()
-    const html = renderToString(sheet.collectStyles(
-      <Header animation={keyframes`0% { opacity: 0; }`} />
-    ))
+    const html = renderToString(
+      sheet.collectStyles(<Header animation={keyframes`0% { opacity: 0; }`} />)
+    )
     const css = sheet.getStyleTags()
 
     expect(html).toMatchSnapshot()
@@ -182,10 +197,12 @@ describe('ssr', () => {
     `
 
     const sheet = new ServerStyleSheet()
-    const html = renderToString(sheet.collectStyles(<Heading>Hello SSR!</Heading>))
+    const html = renderToString(
+      sheet.collectStyles(<Heading>Hello SSR!</Heading>)
+    )
     const elements = sheet.getStyleElement()
 
-    expect(elements).toHaveLength(2);
+    expect(elements).toHaveLength(2)
 
     expect(elements[0].props).toMatchSnapshot()
     expect(elements[1].props).toMatchSnapshot()
@@ -203,12 +220,42 @@ describe('ssr', () => {
     `
 
     const sheet = new ServerStyleSheet()
-    const html = renderToString(sheet.collectStyles(<Heading>Hello SSR!</Heading>))
+    const html = renderToString(
+      sheet.collectStyles(<Heading>Hello SSR!</Heading>)
+    )
     const elements = sheet.getStyleElement()
 
-    expect(elements).toHaveLength(2);
-    expect(elements[0].props.nonce).toBe('foo');
-    expect(elements[1].props.nonce).toBe('foo');
+    expect(elements).toHaveLength(2)
+    expect(elements[0].props.nonce).toBe('foo')
+    expect(elements[1].props.nonce).toBe('foo')
+  })
+
+  it('should interleave styles with rendered HTML when utilitizing streaming', () => {
+    injectGlobal`
+      body { background: papayawhip; }
+    `
+    const Heading = styled.h1`
+      color: red;
+    `
+
+    const sheet = new ServerStyleSheet()
+    const jsx = sheet.collectStyles(<Heading>Hello SSR!</Heading>)
+    const stream = sheet.interleaveWithNodeStream(renderToNodeStream(jsx))
+
+    return new Promise((resolve, reject) => {
+      let received = ''
+
+      stream.on('data', chunk => {
+        received += chunk
+      })
+
+      stream.on('end', () => {
+        expect(received).toMatchSnapshot()
+        resolve()
+      })
+
+      stream.on('error', reject)
+    })
   })
 
   it('should return a generated React style element with minified CSS in production', () => {
@@ -222,10 +269,12 @@ describe('ssr', () => {
     `
 
     const sheet = new ServerStyleSheet()
-    const html = renderToString(sheet.collectStyles(<Heading>Hello Production SSR!</Heading>))
+    const html = renderToString(
+      sheet.collectStyles(<Heading>Hello Production SSR!</Heading>)
+    )
     const elements = sheet.getStyleElement()
 
-    expect(elements).toHaveLength(2);
+    expect(elements).toHaveLength(2)
 
     expect(elements[0].props).toMatchSnapshot()
     expect(elements[1].props).toMatchSnapshot()